aws-scripts/requirements.txt
pandas
requests
pillow
tqdm
pyarrow
boto3
